requirements.txt
InquirerPy~=0.3.4
matplotlib~=3.10.1
mplfinance~=0.12.10b0
pandas~=2.2.3
python-telegram-bot~=21.11.1
requests~=2.32.3
schedule~=1.2.2